+++ packages/lib-react-components/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Project card badges are now circular in Safari.
 - Inconsistent text sizes for Markdown tables and lists.
 - `useUnreadNotifications` and `useUnreadMessages` now wait for Panoptes auth to complete before fetching data.
+- `Modal`: add the `dialog` role and an accessible name. Pass keyboard focus to dialogs when they open.
 
 ## [1.13.0] 2024-05-17